indicate that forgot-password doesn't reset
my password is a randomly generated key in my password manager,
so I prefer to login via the 'forgot pasword' tool. However
I've not really 'Forgotten' my password, and I don't want to
reset it. It took me a while to realise I could still use this
feature as a magic-login-link when starting out.